+++ hydromt/_compat.py
@@ -7,6 +7,9 @@ HAS_GCSFS = False
 HAS_OPENPYXL = False
 HAS_PYET = False
 HAS_S3FS = False
+HAS_BOTO3 = False
+HAS_H5NETCDF = False
+HAS_H5PY = False
 
 try:
     import gcsfs
@@ -40,6 +43,27 @@ try:
     import adlfs
 
     HAS_ADLFS = True
+except ImportError:
+    pass
+
+try:
+    import boto3
+
+    HAS_BOTO3 = True
+except ImportError:
+    pass
+
+try:
+    import h5netcdf
+
+    HAS_H5NETCDF = True
+except ImportError:
+    pass
+
+try:
+    import h5py
+
+    HAS_H5PY = True
 except ImportError:
     pass